+++ server/src-lib/Hasura/Server/Telemetry.hs
@@ -197,7 +197,10 @@ computeActionsMetrics actionCache =
     typeRelationships =
       length . L.nub
         . concatMap
-          (map _trName . maybe [] toList . _otdRelationships . _aotDefinition . snd . _aiOutputObject)
+          ( \aInfo -> map _trName $ case (snd . _aiOutputType $ aInfo) of
+              AOTObject aot -> maybe [] toList . _otdRelationships . _aotDefinition $ aot
+              AOTScalar _ -> []
+          )
         $ actions
 
 -- | Logging related